Various new features
-- Added support for attaching a physical terminal to the console adapter.
When the AUX_TERM_UART_NUM compile-time option is set, the adapter will
relay characters back and forth between the specified Teensy UART device
(default: 2) and the UART connected to the PDP-11 console.  This communication
happens in parallel to the usual flow of data between the host USB serial
device and the console UART.  This makes it possible to have a physical
terminal and/or a host PC connected to the adapter, and to use either to
interact with the PDP-11 console simultaneously.  The serial format and baud
rate for the TTY UART tracks that of the PDP-11 console UART, and is set
automatically based on changes in the host USB serial settings.

-- Add support for 7-E-1 or 7-O-1 serial formats.  The PDP-11/05 CPU is
hard-wired to use the 8-N-1 format.  However, with appropriate software
support, 7-E-1 or 7-O-1 format can be emulated.  To support this, the console
adapter now tracks the serial format configured by the host USB device and
switches the console UART (and the TTY UART) between 8-N-1, 7-E-1 and 7-O-1
as needed.  (All other formats are ignored).  When the console adapter is
booted without a USB connection, the console serial format defaults to 8-N-1.

-- Add support for a test mode that allows convenient loop-back testing of
the console adapter.  By default, when connected to the 11/05 console, the
Teensy’s UART is configured to invert the polarity of serial TX signal.
This matches the polarity expected by the PDP-11 hardware.  However, this
configuration prevents the ability to perform a loopback test the console
adapter by jumpering TX to RX on the console connector.  To make this
possible, the adapter code now supports a test mode in which the TX polarity
is not inverted.  Test mode is entered whenever the test mode pin
(Default: GPIO 4) is grounded when the system boots.  In the console adapter
schematics the test mode pin is connected to an unused pin on the console
connector (TT/38), making it easy to jumper this to one of the ground pins.

-- Updated the schematic to include the RS-232 level shifter circuit (MAX202)
for the physical terminal connection and a jumper to select the adapter power
source.

-- Updated the README