Simplified power circuit and updated documentation
-- Eliminated the jumper used to select between USB and PDP-11 power in favor
of always sourcing power from the PDP.  With this change, one must now use a
wire jumper between VUSB and pin 3 on the console connector to power the
adapter from the USB host.  This simplifies construction and has the added
benefit of making it essentially impossible to mistakenly connect USB +5V to
the PDP-11’s +5V line.

-- Updated the README to reflect the new power circuit design. Added new
sections describing how to modify the Teensy for use with external power and
how to power the adapter from USB.

-- Revised all the README images to show the latest example design.

-- Added README section on building the aux terminal cable.

-- Rearranged the order of sections for clarity and made various other
editorial changes.